re-structure the project and add tests
